Renamed RecipeCollection - updated path in App
