+++ frontend/tests/pages/MyRecipesPage.test.jsx
@@ -1,7 +1,7 @@
 import { render, screen} from "@testing-library/react";
 import { vi, expect, describe, test } from "vitest";
 import { userEvent } from "@testing-library/user-event";
-import { RecipeCollection } from "../../src/pages/RecipeCollection/RecipeCollection";
+import { RecipeCollection } from "../../src/pages/MyRecipes/MyRecipesPage";
 import { useNavigate } from "react-router-dom";
 import { useFetchRecipes } from "../../src/hooks/useFetchRecipe";
